--- frontend/src/routes/pages/Join.tsx
@@ -1,5 +1,0 @@
-function Join() {
-    return <></>
-}
-
-export default Join;
